+++ lib/vault_db/core/daos/base/otp/otp_items_dao.dart
@@ -1,6 +1,6 @@
 import 'package:drift/drift.dart';
-
 import 'package:hoplixi/vault_db/core/vault_db.dart';
+
 import '../../../scheme/tables/otp/otp_items.dart';
 
 part 'otp_items_dao.g.dart';